requirements.txt
requests
beautifulsoup4

# Core HTTP
requests>=2.31.0
urllib3>=2.0.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0

# CLI / Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Google Maps Enrichment (if enabled)
googlemaps>=4.10.0